MTH229g

Documentation for MTH229g.

MTH229g.MTH229g

MTH229g.bisection

MTH229g.fisheye

MTH229g.fubini

MTH229g.newton

MTH229g.MTH229g — Module

MTH229: helper functions for using Julia with MTH229.

This package reexports functions from the CalculusWithJulia package.

Use mth229() to download course notebooks.

MTH229g.bisection — Method

Simple implementation of the bisection method.

Example:

bisection(sin, 3, 4)
f(x) = x^5 - x^4 - x^3 - x^2 - x - 1
a = bisection(f, 1, 2)
f(a)

The display shows a simple graphic illustrating the method's division for the first few steps.

An easier-to-understand alternative to Roots.find_zero(f, (a,b), Bisection()).

MTH229g.fisheye — Method

fisheye(f)

Transform f defined on (-∞, ∞) to a new function whose domain is in (-π/2, π/2) and range is within (-π/2, π/2). Useful for finding all zeros over the real line. For example

f(x) = 1 + 100x^2 - x^3
fzeros(f, -100, 100) # empty just misses the zero found with:
fzeros(fisheye(f), -pi/2, pi/2) .|> tan  # finds 100.19469143521222, not perfect but easy to get

By Gunter Fuchs.

MTH229g.fubini — Method

fubini(f, [zs], [ys], xs; rtol=missing, kws...)

Integrate f of 1, 2, or 3 input variables.

The zs may depend (x,y), the ys may depend on x

Examples

# integrate over the unit square
fubini((x,y) -> sin(x-y), (0,1), (0,1))

# integrate over a triangle
fubini((x,y) -> 1, (0,identity), (0,1 ))

#
f(x,y,z) = x*y^2*z^3
fubini(f, (0,(x,y) ->  x+ y), (0, x -> x), (0,1))

!!! Note This uses nested calls to quadgk. The use of hcubature is recommended, typically after a change of variables to make a rectangular domain. The relative tolerance increases at each nested level.

MTH229g.newton — Method

newton(f, [fp], x0; verbose=false, kwargs...)

Run Newton's method to find a zero of f near x0.